Put assign mapping to a separate folder
